Added basic config handling.
Still work to do on this before it's really workable for all the settings required, but this at least gets things working well. It uses an MIT licensed library for simple INI reading and supports writing, so GUI options can later be added for writing the config.

Currently only supports the fullscreen option (see the included config.ini) which allows you to set the initial state of the application. The config code should be pulled off into its own function for long-ter
m cleanliness, but I wanted to get a pull request together and confer with the maintainer regarding style decisions there.