Use a Nix Flake for m8c
Flakes are a different way to specify a Nix project/configuration. It
provides a few benefits over the existing solution:

  1. The exact versions of inputs are snapshot. This should mean that a
     build on one machine will use the exact same packages as the build
     on another.
  2. Flakes can be consumed by other Flakes. This would mean that folks
     could use this Flake as an import and then declaritively configure
     their `m8c` setup.
  3. While not used here now, this could make versioning easier. There
     wouldn't need to be a `fetchFromGitHub` step because someone could
     just use/run the app for that tagged branch. Ex: `nix run
     github:laamaa/m8c?ref=v1.7.7`

That said, this PR does rely on flake-compat to provide backward
compatibility with `default.nix` and `shell.nix`. So a user could use
`nix-shell` as before, or they could use `nix develop` (and potentially
devenv in the future).