Update nixpkgs so that the SDL3 dependency can be found
This migrates to nixpkgs-unstable for always getting the latest
dependencies (provided `nix flake update` is run). As a result, SDL3 can
be found, though the dependency is spelled with lower-case instead of
upper case.

The cmake package was added to the dev shell so that a MacOS build could
also be tested with these changes.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>