Simplify the Nix config and add a workflow for it
Got rid of m8c-stable and made m8c-dev the m8c candidate. If ever a user
wants a specific version, they need only use the git tags to get that
specific version. This reduces the complexity of keeping track of hashes
inside of the flake since it always just builds to what code is checked
out.

In simplifying, updated the flake to use CMake, like the workflows for Mac
and Linux do. This should keep things a bit more consistent that way and
reduce one more case of depdending on the Makefile. In order to use
CMake though, I had to add some install rules for the binary. While I
was there I also put a few extra files into appropriate directories -
like documents and the game controller DB file.

Lastly, to make this feel less of a headache for folks, I added a new
job to the ubuntu workflow to build out the project with Nix.

Signed-off-by: K. Adam Christensen <pope@shifteleven.com>